SET 1

+++ InstaStuff/InstaStuff/Classes/Managers/TemplatesStorage/TemplateSet.swift
@@ -10,7 +10,7 @@ import UIKit
 
 struct TemplateSet {
     
-    typealias TemplateSetId = Int
+    typealias TemplateSetId = Int64
     
     // MARK: - Properties
